Fully centered table header
- Removed empty space from grid headers, which was taken by the column's
sort arrow which most of the time is not used at all by the users.
Moreover column's without sorting enabled were not taking the the space
at all, which resulted in inconsistent header alignment in different
grid columns.
- Equalized left padding and right padding in grid headers, because
right padding was most of the taken by the column's filter, resulting in
not fully centered header titles, which confuses the client into
thinking that header titles are aligned randomly.